sns_backend/sns_backend_rest/nsfw/requirements.txt
appdirs==1.4.3
cycler==0.10.0
decorator==4.0.11
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.6.2
h5py==2.7.0
httpie==0.9.9
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
Keras==2.0.4
matplotlib==2.0.2
numpy==1.12.1
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorflow==1.1.0
Theano==0.9.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
